server/requirements.txt
Flask==0.10.1
Flask-Cors==2.0.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
argparse==1.2.1
gevent==1.0.2
greenlet==0.4.7
itsdangerous==0.24
pyOpenSSL==0.13
pycrypto==2.6.1
requests==2.7.0
six==1.9.0
soco==0.11.1
wsgiref==0.1.2
